@@ -4,7 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "eslint": "eslint ./src index.js --fix"
+    "start": "ts-node index.ts",
+    "lint": "eslint ./src/**/*.ts index.ts"
   },
   "author": "",
   "license": "MIT",
@@ -13,6 +14,12 @@
     "mariadb": "^2.5.4"
   },
   "devDependencies": {
-    "eslint": "^7.28.0"
+    "@types/node": "^16.6.1",
+    "@typescript-eslint/eslint-plugin": "^4.29.1",
+    "@typescript-eslint/parser": "^4.29.1",
+    "eslint": "^7.32.0",
+    "eslint-plugin-no-else": "^0.2.2",
+    "ts-node": "^10.1.0",
+    "typescript": "^4.3.5"
   }
 }
