@@ -2,8 +2,10 @@ module.exports = {
     apps : [{
         name: 'eve',
         cwd: './',
-        script: './index.js',
+        script: './node_modules/ts-node/dist/bin.js',
+        args: ['./index.ts'],
         watch: ['../'],
+        autorestart: false,
         max_restarts: 1,
         restart_delay: 5000,
         env: {
